added interactive authorization
I readded some of the basic structure of gocjdroute, starting from
scratch and borrowing some routines. All that the main function
currently does is force the appending of a password block to
example.conf.

The Authorize() function has become a sort of all-in-one, though it
currently only understands user input. It should be modified slightly to
accept pre-parsed authorization blocks. It also uses encoding/json to
accurately display what is going into the configuration file. (It's made
to be copy-pastable into a configuration file, as well, in case
GoCjdroute is used for a dry-run.)