added flags and UI syntax
The basic syntax is "gocjdroute subcommand arguments" where subcommand
is something like "auth" "conn" or "remove", though only "auth" is
implemented currently. The "arguments" area is parsed based on the
command, or replaced by a simple default value.

This also adds a framework for adding and implementing further commands
and flags.