added easy configuration generation
Authorize() now generates a copy-pastable Connection block. It is
generated on the fly, asking for and saving information to identify you
to your peer, including name and location, as well as the (required)
IPv4 connection details. These are saved back to the configuration along
with the new authorization block, and are only asked for if they are not
provided in the configuration