fixed JSON ambiguity
I adjusted the way that GoCjdroute handles arguments when --json is
enabled, and causes it to take implicit values in that case. The
syntaxes are "gocjdroute auth [index] <json>", where index, if not
included, is -1, (the JSON is parsed properly whether it is included or
not,) and "gocjdroute conn [discarded] <json>", where the middle value
can be included or not, but is not used.

closes #10