adjusted input and output styles
I removed the use of the fmt package, because it was relatively
unnecessary. Furthermore, I fixed the problem with multi-word input. I
scrapped the fmt.Scanln() function, which wasn't working properly, and
instead used bufio and os.Stdin to create an equivalent.

The newlines on senseless systems, such as Windows, may break this,
however. It may bear further testing.

closes #13